Admin manage issues

+++ src/pages/AdminUserManager.tsx
@@ -5,21 +5,23 @@ import {useEffect, useState} from "react";
 import axios from "axios";
 import {useNavigate} from "react-router-dom";
 import {toast} from "react-hot-toast";
+import TableList from "../components/TableList.tsx";
 
 const AdminUserManager = () => {
-    const user = useAuth([AccountType.Admin]);
+    useAuth([AccountType.Admin]);
 
     const navigate = useNavigate();
 
-    const [users, setUsers] = useState<Array<User>>([])
-    const [selectedUsers, setSelectedUsers] = useState([])
+    const [users, setUsers] = useState<Array<User>>([]);
+
+    const [data, setData] = useState([]);
 
     useEffect(() => {
         axios.get(SERVER_URL + "/api/users", RequestConfig())
             .then(response => {
             setUsers(response.data);
-        })
-    }, [])
+        });
+    }, []);
 
     const getType = (t: number) => {
         switch (t) {
@@ -30,36 +32,46 @@ const AdminUserManager = () => {
                 return "Employee";
 
             case 2:
-                return "Admin"
+                return "Admin";
 
             default:
-                return "None"
-        }
-    }
-
-    const handleSelect = (obj: EventTarget & HTMLInputElement) => {
-        const id = parseInt(obj.name);
-
-        if (selectedUsers.includes(id)) {
-            setSelectedUsers([...selectedUsers.filter(i => i !== id)]);
-        } else {
-            setSelectedUsers([...selectedUsers.concat([id])]);
+                return "None";
         }
     };
 
-    const handleDelete = () => {
-        const requests = selectedUsers.map(userId => {
-            axios.delete(SERVER_URL + `/api/users/${userId}`, RequestConfig());
+    useEffect(() => {
+        const _data = [];
+        users.forEach(u => {
+            _data.push([
+                u.id,
+                u.username,
+                getType(u.type),
+                u.phoneNumber,
+                u.unit
+            ]);
         });
 
-        toast.promise(Promise.all(requests), {
+        setData(_data);
+    }, [users]);
+
+    const handleDelete = (userId) => {
+        if (!window.confirm("Are you sure that you want to delete this user?")) {
+            return;
+        }
+
+        const request = axios.delete(SERVER_URL + `/api/users/${userId}`, RequestConfig());
+
+        toast.promise(request, {
             loading: "Loading...",
-            success: `Deleted ${requests.length} ${requests.length == 1 ? "user" : "users"}`,
-            error: "Failed to delete users"
+            success: "Deleted user",
+            error: "Failed to delete user"
         }).then(() => {
-            setUsers([...users.filter(u => !selectedUsers.includes(u.id))])
-            setSelectedUsers([]);
-        })
+            setUsers([...users.filter(u => u.id !== userId)]);
+        });
+    };
+
+    const handleEdit = (userId) => {
+        navigate(`/admin/users/edit/${userId}`);
     };
 
     const handleNewUser = () => {
@@ -81,43 +93,28 @@ const AdminUserManager = () => {
     return <>
         <NavigationHeader/>
         <div className={"page-content"}>
-            <div className={"admin-users-edit-header"}>
-                <h1>Users</h1>
-                <button className={"admin-users-add-user"} onClick={handleNewUser}><i className="fa-solid fa-user-plus"></i></button>
-            </div>
-            <div className={"admin-users-edit-header"}>
-                {selectedUsers.length > 0
-                ? <>
-                        <p>{selectedUsers.length} {selectedUsers.length == 1 ? "item" : "items"} selected</p>
-                        <a onClick={handleDelete}>Delete</a>
-                    </>
-                : <></>}
-            </div>
-            <div className={"admin-users-legenda"}>
-                <p></p>
-                <p></p>
-                <p>Id</p>
-                <p>Username</p>
-                <p>Type</p>
-                <p>Phone number</p>
-                <p>Unit</p>
-            </div>
-            <div className={"admin-users-list"}>
-                {users.map(u => (
-                    <div className={"admin-users-list-item"} key={u.id}>
-                        <input type={"checkbox"} name={u.id.toString()} onChange={(e) => handleSelect(e.target)}/>
-                        <p onClick={() => navigate(`edit/${u.id}`)}>
-                            <i className="fa-solid fa-pen-to-square"></i>
-                        </p>
-                        <p>{u.id}</p>
-                        <p>{u.username}</p>
-                        <p>{getType(u.type)}</p>
-                        <p>{u.phoneNumber}</p>
-                        <p>{u.unit}</p>
-                    </div>))}
-            </div>
+            <h1>Users</h1>
+            <button onClick={handleNewUser}
+                    style={{
+                        width: "100px",
+                        borderRadius: "4px"
+                    }}>
+                Add user <i className="fa-solid fa-user-plus"></i>
+            </button>
+            <TableList columns={["ID", "Username", "Type", "Phone number", "unit"]}
+                       data={data}
+                       buttons={[
+                           {
+                               text: <i className="fa-solid fa-pen-to-square"></i>,
+                               callback: handleEdit
+                           },
+                           {
+                               text: <i className="fa-solid fa-trash"></i>,
+                               callback: handleDelete
+                           }
+                       ]}/>
         </div>
-    </>
-}
+    </>;
+};
 
 export default AdminUserManager;
